refactor: moved hasPostions to hook file
